Add concrete FlightRecorderDiskSource
