+++ authenticator/src/main/kotlin/com/bitwarden/authenticator/data/platform/repository/SettingsRepository.kt
@@ -2,6 +2,7 @@ package com.bitwarden.authenticator.data.platform.repository
 
 import com.bitwarden.authenticator.ui.platform.feature.settings.appearance.model.AppLanguage
 import com.bitwarden.authenticator.ui.platform.feature.settings.data.model.DefaultSaveOption
+import com.bitwarden.data.manager.flightrecorder.FlightRecorderManager
 import com.bitwarden.ui.platform.feature.settings.appearance.model.AppTheme
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
@@ -9,7 +10,7 @@ import kotlinx.coroutines.flow.StateFlow
 /**
  * Provides an API for observing and modifying settings state.
  */
-interface SettingsRepository {
+interface SettingsRepository : FlightRecorderManager {
 
     /**
      * The [AppLanguage] for the current user.
