+++ authenticator/src/main/kotlin/com/bitwarden/authenticator/ui/platform/manager/di/PlatformUiManagerModule.kt
@@ -1,6 +1,8 @@
 package com.bitwarden.authenticator.ui.platform.manager.di
 
+import com.bitwarden.authenticator.ui.platform.manager.AuthenticatorBuildInfoManagerImpl
 import com.bitwarden.authenticator.ui.platform.model.SnackbarRelay
+import com.bitwarden.core.data.manager.BuildInfoManager
 import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
 import com.bitwarden.ui.platform.manager.snackbar.SnackbarRelayManager
 import com.bitwarden.ui.platform.manager.snackbar.SnackbarRelayManagerImpl
@@ -16,6 +18,10 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 class PlatformUiManagerModule {
+    @Provides
+    @Singleton
+    fun provideBuildInfoManager(): BuildInfoManager = AuthenticatorBuildInfoManagerImpl()
+
     @Provides
     @Singleton
     fun provideSnackbarRelayManager(
