+++ authenticator/src/test/kotlin/com/bitwarden/authenticator/data/platform/repository/SettingsRepositoryTest.kt
@@ -23,8 +23,9 @@ class SettingsRepositoryTest {
         every { getAlertThresholdSeconds() } returns 7
     }
 
-    private val settingsRepository = SettingsRepositoryImpl(
+    private val settingsRepository: SettingsRepository = SettingsRepositoryImpl(
         settingsDiskSource = settingsDiskSource,
+        flightRecorderManager = mockk(),
         dispatcherManager = FakeDispatcherManager(),
     )
