+++ data/src/main/kotlin/com/bitwarden/data/manager/di/DataManagerModule.kt
@@ -3,12 +3,15 @@ package com.bitwarden.data.manager.di
 import android.content.Context
 import com.bitwarden.core.data.manager.BuildInfoManager
 import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
+import com.bitwarden.data.datasource.disk.FlightRecorderDiskSource
 import com.bitwarden.data.manager.BitwardenPackageManager
 import com.bitwarden.data.manager.BitwardenPackageManagerImpl
 import com.bitwarden.data.manager.NativeLibraryManager
 import com.bitwarden.data.manager.NativeLibraryManagerImpl
 import com.bitwarden.data.manager.file.FileManager
 import com.bitwarden.data.manager.file.FileManagerImpl
+import com.bitwarden.data.manager.flightrecorder.FlightRecorderManager
+import com.bitwarden.data.manager.flightrecorder.FlightRecorderManagerImpl
 import com.bitwarden.data.manager.flightrecorder.FlightRecorderWriter
 import com.bitwarden.data.manager.flightrecorder.FlightRecorderWriterImpl
 import com.bitwarden.network.service.DownloadService
@@ -43,6 +46,22 @@ object DataManagerModule {
         context = context,
         downloadService = downloadService,
         dispatcherManager = dispatcherManager,
+    )
+
+    @Provides
+    @Singleton
+    fun provideFlightRecorderManager(
+        @ApplicationContext context: Context,
+        clock: Clock,
+        dispatcherManager: DispatcherManager,
+        flightRecorderDiskSource: FlightRecorderDiskSource,
+        flightRecorderWriter: FlightRecorderWriter,
+    ): FlightRecorderManager = FlightRecorderManagerImpl(
+        context = context,
+        clock = clock,
+        dispatcherManager = dispatcherManager,
+        flightRecorderDiskSource = flightRecorderDiskSource,
+        flightRecorderWriter = flightRecorderWriter,
     )
 
     @Provides
