+++ data/src/main/kotlin/com/bitwarden/data/manager/flightrecorder/FlightRecorderManager.kt
@@ -1,12 +1,8 @@
 package com.bitwarden.data.manager.flightrecorder
 
-import android.content.Context
-import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
-import com.bitwarden.data.datasource.disk.FlightRecorderDiskSource
 import com.bitwarden.data.datasource.disk.model.FlightRecorderDataSet
 import com.bitwarden.data.manager.model.FlightRecorderDuration
 import kotlinx.coroutines.flow.StateFlow
-import java.time.Clock
 
 /**
  * Manager class that handles recording logs for the flight recorder.
@@ -46,24 +42,4 @@ interface FlightRecorderManager {
      * Deletes the raw log files and metadata.
      */
     fun deleteAllLogs()
-
-    @Suppress("UndocumentedPublicClass")
-    companion object {
-        /**
-         * Creates a new instance of the [FlightRecorderManager].
-         */
-        fun create(
-            context: Context,
-            clock: Clock,
-            flightRecorderDiskSource: FlightRecorderDiskSource,
-            flightRecorderWriter: FlightRecorderWriter,
-            dispatcherManager: DispatcherManager,
-        ): FlightRecorderManager = FlightRecorderManagerImpl(
-            context = context,
-            clock = clock,
-            flightRecorderDiskSource = flightRecorderDiskSource,
-            flightRecorderWriter = flightRecorderWriter,
-            dispatcherManager = dispatcherManager,
-        )
-    }
 }
